good ms

+++ src/main/java/ch/heig/gre/labo2/groupA/Recorder.java
@@ -33,7 +33,7 @@ public class Recorder {
         if (_startTimeStamp == -1) {
             return null;
         }
-        RecorderOutput output = new RecorderOutput(_nVertexFromFIFO, _nEdgeCompute, _nRelaxation, _nVertextMissing, (System.nanoTime() - _startTimeStamp) / 1000000);
+        RecorderOutput output = new RecorderOutput(_nVertexFromFIFO, _nEdgeCompute, _nRelaxation, _nVertextMissing, (double) (System.nanoTime() - _startTimeStamp) / 1000000);
         _startTimeStamp = -1;
         return output;
     }
